feat(web): editor de mapas y panel al día con coordenadas globales
Adapta el editor de mapas del panel de administración (exportadores de
puertas y spawns, historial, selección y redimensionado) al mundo global,
y documenta la tubería de mapa en vivo.

+++ src/Web/Map/GameMapInfoAdapter.cs
@@ -120,8 +120,8 @@ internal sealed class GameMapInfoAdapter : IGameMapInfo, IDisposable
 
         public string AccountName => this._player.Account?.LoginName ?? "N/A";
 
-        public byte LocationX => this._player.Position.X;
+        public ushort LocationX => this._player.Position.X;
 
-        public byte LocationY => this._player.Position.Y;
+        public ushort LocationY => this._player.Position.Y;
     }
 }
